+++ app/__init__.py
@@ -1,32 +1,18 @@
-from flask import Flask, request
+from flask import Flask
 import config
-from app import search
-from flask_pymongo import PyMongo
+from app import search, get_detail_using_gid
+from app.views.api import api
+# from flask_pymongo import PyMongo
 
 
 def create_app(config_file=config.Config):
     app = Flask(__name__)
     app.config.from_object(config_file)
-    mongo = PyMongo(app)
+    # mongo = PyMongo(app)
+    app.register_blueprint(api)
 
     @app.route('/')
     def hello_world():
         return 'Hello World!'
-
-
-    @app.route('/api/searchByWords')
-    def search_by_words():
-        """ 主要搜索API """
-        words = request.args.get('wd')
-        #: (optional) 搜索的关键字，如果为空则返回所有结果
-        limit = request.args.get('lim')
-        #: (optional) 搜索的结果数，默认25
-        page = request.args.get('pg')
-        #: (optional) 第几页的内容，默认1
-        filtered_category = request.args.get('fc')
-        #: (optional) 不包含的分类
-        minimum_rating = request.args.get('mr')
-        #: (optional) 最低评分
-        return search.search(words, limit, page, filtered_category, minimum_rating)
 
     return app
