+++ app/get_detail_using_gid.py
@@ -1,0 +1,2 @@
+def get_detail_using_gid(gid):
+    return 'Not Implemented'
