+++ app/views/api.py
@@ -1,0 +1,26 @@
+from flask import Blueprint, request
+from app import search, get_detail_using_gid
+
+api = Blueprint('api', __name__)
+
+
+@api.route('/api/searchByWords')
+def search_by_words():
+    """ 主要搜索API """
+    words = request.args.get('wd')
+    #: (optional) 搜索的关键字，如果为空则返回所有结果
+    limit = request.args.get('lim')
+    #: (optional) 搜索的结果数，默认25
+    page = request.args.get('pg')
+    #: (optional) 第几页的内容，默认1
+    filtered_category = request.args.get('fc')
+    #: (optional) 不包含的分类
+    minimum_rating = request.args.get('mr')
+    #: (optional) 最低评分
+    return search.search(words, limit, page, filtered_category, minimum_rating)
+
+
+@api.route('/api/getDetail')
+def get_detail():
+    gid = request.args.get('gid')
+    return get_detail_using_gid.get_detail_using_gid(gid)
